requirements.txt
absl-py>=2.2.1
accelerate>=1.6.0
addict>=2.4.0
asttokens>=3.0.0
certifi>=2025.1.31
charset-normalizer>=3.4.1
contourpy>=1.3.1
cycler>=0.12.1
decorator>=5.2.1
defusedxml>=0.7.1
diffusers>=0.32.2
exceptiongroup>=1.2.2
executing>=2.2.0
filelock>=3.18.0
fonttools>=4.56.0
fsspec>=2025.3.2
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git@856dde20aee659246248e20734ef9ba5214f5e44
grpcio>=1.71.0
huggingface-hub>=0.30.1
idna>=3.10
importlib_metadata>=8.6.1
ipython>=8.34.0
jedi>=0.19.2
Jinja2>=3.1.6
kiwisolver>=1.4.8
Markdown>=3.7
MarkupSafe>=3.0.2
matplotlib>=3.10.1
matplotlib-inline>=0.1.7
mpmath>=1.3.0
networkx>=3.4.2
numpy>=2.2.4
nvidia-cublas-cu12>=12.1.3.1
nvidia-cuda-cupti-cu12>=12.1.105
nvidia-cuda-nvrtc-cu12>=12.1.105
nvidia-cuda-runtime-cu12>=12.1.105
nvidia-cudnn-cu12>=9.1.0.70
nvidia-cufft-cu12>=11.0.2.54
nvidia-curand-cu12>=10.3.2.106
nvidia-cusolver-cu12>=11.4.5.107
nvidia-cusparse-cu12>=12.1.0.106
nvidia-cusparselt-cu12>=0.6.2
nvidia-nccl-cu12>=2.21.5
nvidia-nvjitlink-cu12>=12.4.127
nvidia-nvtx-cu12>=12.1.105
opencv-python>=4.11.0.86
packaging>=24.2
pandas>=2.2.3
parso>=0.8.4
pexpect>=4.9.0
pillow>=11.2.0
platformdirs>=4.3.7
prompt_toolkit>=3.0.50
protobuf>=6.30.2
psutil>=7.0.0
ptyprocess>=0.7.0
pure_eval>=0.2.3
pycocotools>=2.0.8
Pygments>=2.19.1
pyparsing>=3.2.3
python-dateutil>=2.9.0.post0
pytz>=2025.2
PyYAML>=6.0.2
regex>=2024.11.6
requests>=2.32.3
safetensors>=0.5.3
scipy>=1.15.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
six>=1.17.0
stack-data>=0.6.3
supervision>=0.25.1
sympy>=1.13.1
tensorboard>=2.19.0
tensorboard-data-server>=0.7.2
timm>=1.0.15
tokenizers>=0.21.1
tomli>=2.2.1
torch>=2.5.1+cu121
torchvision>=0.20.1+cu121
tqdm>=4.67.1
traitlets>=5.14.3
transformers>=4.50.3
triton>=3.1.0
typing_extensions>=4.13.0
tzdata>=2025.2
urllib3>=2.3.0
wcwidth>=0.2.13
Werkzeug>=3.1.3
yapf>=0.43.0
zipp>=3.21.0
